refactor(KERN): remove unused userFeedback components
