+++ app/components/content/ContentComponents.tsx
@@ -12,9 +12,9 @@ import KernRichText from "../kern/KernRichText";
 import KernHeading from "../kern/KernHeading";
 import { KernInlineNotice } from "../kern/KernInlineNotice";
 import KernVideo from "../kern/video/KernVideo";
-import KernUserFeedback from "../kern/UserFeedback";
 import { KernEmailCapture } from "~/components/kern/emailCapture/KernEmailCapture";
 import { KernDetails } from "../kern/KernDetails";
+import KernUserFeedback from "./userFeedback";
 
 function getContentBackgroundColor(el: StrapiContentComponent): string {
   if ("contentBackgroundColor" in el) {
