--- app/components/content/userFeedback/FeedbackType.tsx
@@ -1,1 +1,0 @@
-export type FeedbackType = "positive" | "negative";
